update: Emailing service completed, working on email styling

+++ backend/models/user.py
@@ -13,10 +13,10 @@ class User(base):
     email = Column(String(255), index = True)
     password = Column(String(255), index = True)
     subscribed = Column((Boolean),index=True)
-    property_city = Column(Integer(10), index = True)
-    property_type = Column(Integer(10), index = True)
-    min_price = Column(Integer(20), index = True)
-    max_price = Column(Integer(20), index = True)
-    property_sqm = Column(Integer(20), index = True)
+    property_city = Column(Integer, index = True)
+    property_type = Column(Integer, index = True)
+    min_price = Column(Integer, index = True)
+    max_price = Column(Integer, index = True)
+    property_sqm = Column(Integer, index = True)
